backend/requirements.txt
fastapi[all]
uvicorn
python-dotenv
supabase
python-jose[cryptography]
python-multipart
pdfplumber
python-docx
pytesseract
opencv-python
transformers
fairlearn
torch 
pydantic

